Move plural test to experimental
